+++ src/test/java/drawie/SocketTests.java
@@ -1,0 +1,66 @@
+/*
+package drawie;
+
+import io.socket.client.Socket;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.mockito.*;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.when;
+
+
+public class SocketTests {
+
+    private String roomUrl = "https://drawie.herokuapp.com/?room=3332b08b-a9f5-4c40-32s21c-9ea26a3c5ef1";
+
+    @Mock
+    Model model;
+
+    */
+/*    @Mock*//*
+
+
+*/
+/*
+    @Mock
+    private Socket socket;*//*
+
+*/
+/*
+    @Before
+    public void configureSocket(){
+        socket = mySocket.getSocket();
+    }*//*
+
+
+*/
+/*
+
+
+    @InjectMocks
+    MySocket mySocket;
+
+    @Mock
+    Socket socket;
+
+    @BeforeEach
+    public void init() {
+        MockitoAnnotations.initMocks(this);
+        System.out.println("OOO");
+    }
+
+    @Test
+    public void whenRedoIsEmittedSocketEmits() {
+
+        when(socket.emit("redo")).thenReturn(null);
+        System.out.println(socket.id());
+        assertEquals(mySocket.emit("redo"), 2);
+        //mySocket.emit("redo");
+        //Mockito.verify(socket.emit("redo"), Mockito.atLeastOnce());
+    }
+
+*//*
+
+}
+*/
